Trier les catégories de danger par ordre alpha
- Sur les pages de création / modification
- Dans le filtre sur la vue de liste
