+++ .talismanrc
@@ -43,3 +43,5 @@ fileignoreconfig:
   checksum: ad80322fb2dab542fe667054b83761da8e386ce7357b123b2e30a0fea28f1f10
 - filename: .github/workflows/django.yml
   checksum: bb2812e864802e3e55d05b28f6da146973a78b382dad3d94d1de8d4f10430657
+- filename: ssa/models/mixins.py
+  checksum: d3c07c3669e197136987b8fc9d4d0a837484e5976455aa31e782032d7964802b
